service-workers/service-worker/resource-timing-bodySize-cross-origin.https.html

<!DOCTYPE html>
<meta charset="utf-8">
<script src="/common/utils.js"></script>
<script src="/common/get-host-info.sub.js"></script>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/test-helpers.sub.js"></script>
<script>
const {REMOTE_ORIGIN} = get_host_info();

function test_scenario({tao, mode}) {
    promise_test(async (t) => {
        const uid = token();
        const worker_url = `resources/passthrough.js?uid=${uid}`;
        const scope = `resources/fetch-response.html?uid=${uid}`;
        const iframe = document.createElement('iframe');
        iframe.src = `${scope}&path=${encodeURIComponent(`${REMOTE_ORIGIN}/element-timing/resources/TAOImage.py?origin=*&tao=${tao === "pass" ? "wildcard" : "none"}`)}&mode=${mode}`;
        const registration = await service_worker_unregister_and_register(t, worker_url, scope);
        t.add_cleanup(() => registration.unregister());
        t.add_cleanup(() => iframe.remove());
        await wait_for_state(t, registration.installing, 'activated');
        const waitForMessage = new Promise(resolve => window.addEventListener('message', ({data}) => resolve(data)));
        document.body.appendChild(iframe);
        const {buffer, entry} = await waitForMessage;
        if (mode === "cors")
          assert_greater_than(buffer.byteLength, 0);
        else
          assert_equals(buffer.byteLength, 0);
        assert_equals(entry.decodedBodySize, buffer.byteLength);
        assert_equals(entry.encodedBodySize, buffer.byteLength);
        assert_equals(entry.encodedBodySize, entry.decodedBodySize);
        if (mode === "cors")
            assert_greater_than(entry.decodedBodySize, 0);
    }, `Service-worker body size: TAO ${tao} ${mode}`);
}

for (const mode of ["cors", "no-cors"]) {
  for (const tao of ["pass", "fail"])
    test_scenario({tao, mode});
}

</script>
